requirements.txt
asgiref==3.2.7
celery==4.4.5
channels==2.4.0
channels-redis==2.4.2
Django==3.0.5
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
pytz==2019.3
sqlparse==0.3.1
